+++ apps/docs/docusaurus.config.js
@@ -41,7 +41,7 @@ const config = {
           routeBasePath: '/',
           versions: {
             current: {
-              label: '1.0.0-alpha.1',
+              label: '1.0.0-alpha.0',
               path: '',
             },
           },