+++ apps/docs/src/theme/NavbarItem/ThemeDropdownNavbarItem.js
@@ -1,19 +1,22 @@
 import BrowserOnly from '@docusaurus/BrowserOnly';
+import useIsBrowser from '@docusaurus/useIsBrowser';
+import { THEME_DROPDOWN_DEFAULT_SELECTED } from '@site/src/utils/constants';
 import React, { useEffect, useState } from 'react';
+import {useLocation} from '@docusaurus/router';
 
-const THEME_DROPDOWN_DEFAULT_SELECTED = 2;
-
-const mapItems = (options) =>
-  options.map((opt, idx) => {
+const mapItems = (options, defaultSelected) =>
+  options.map((opt) => {
     return {
       label: opt.label,
       value: opt.label,
       // preserve ?search#hash suffix on version switches
       // to: `${versionDoc.path}${search}${hash}`,
       isActive: () => {
-        if (idx === THEME_DROPDOWN_DEFAULT_SELECTED) {
+        if (opt.label === defaultSelected) {
           return true;
         }
+
+        return false;
       },
       onClick: () => {},
     };
@@ -26,25 +29,36 @@ export default function ThemeDropdownNavbarItem({
   defaultSelected,
   ...props
 }) {
-  const items = [...mapItems(options)];
+  const isBrowser = useIsBrowser();
+  const location = useLocation();
+  const items = [...mapItems(options, defaultSelected)]
 
-  const [activeTheme, setActiveTheme] = useState(items.find((x) => x.isActive()));
+  const [activeTheme, setActiveTheme] = useState(
+    items.find((x) => x.isActive())
+  );
 
   useEffect(() => {
     const body = document.querySelector('body');
-    body.classList.add(activeTheme.value.replaceAll(" ", "-").toLowerCase());
-  }, []);
+    if(!body.classList.contains(activeTheme.value.replaceAll(' ', '-').toLowerCase())) {
+      body.classList.add(activeTheme.value.replaceAll(' ', '-').toLowerCase());
+    }
+  }, [location])
 
   const _selectHandler = (event) => {
     const body = document.querySelector('body');
-    const theme = event.detail?.value.replaceAll(" ", "-").toLowerCase();
+    const theme = event.detail?.value.replaceAll(' ', '-').toLowerCase();
     setActiveTheme((prev) => {
-      if(body.classList.contains(prev.value.replaceAll(" ", "-").toLowerCase())) {
-        body.classList.remove(prev.value.replaceAll(" ", "-").toLowerCase());
+      if (
+        body.classList.contains(prev.value.replaceAll(' ', '-').toLowerCase())
+      ) {
+        body.classList.remove(prev.value.replaceAll(' ', '-').toLowerCase());
       }
       body.classList.add(theme);
       return event.detail;
-    })
+    });
+    if (isBrowser) {
+      window.localStorage.setItem('activeTheme', event.detail.value)
+    };
   };
 
   const _items =
@@ -55,17 +69,18 @@ export default function ThemeDropdownNavbarItem({
 
   return (
     <BrowserOnly>
-    {() => {
-    const StudsDropdown =
-      require('@studs/react').StudsDropdown;
-      return <StudsDropdown
-      class="theme-dropdown"
-      size="small"
-      selected={activeTheme.value}
-      onChange={_selectHandler}
-      options={_items}
-    ></StudsDropdown>;
-  }}
-  </BrowserOnly>
+      {() => {
+        const StudsDropdown = require('@studs/react').StudsDropdown;
+        return (
+          <StudsDropdown
+            class="theme-dropdown"
+            size="small"
+            selected={activeTheme.value}
+            onChange={_selectHandler}
+            options={_items}
+          ></StudsDropdown>
+        );
+      }}
+    </BrowserOnly>
   );
 }