+++ apps/docs/src/utils/constants.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import ExecutionEnvironment from '@docusaurus/ExecutionEnvironment';
 
 export const FORM_PROPS = [
   {
@@ -99,3 +100,6 @@ export const THEME_DROPDOWN_OPTIONS = [
     label: 'Supplier Tools',
   },
 ];
+
+export const THEME_DROPDOWN_DEFAULT_SELECTED = ExecutionEnvironment.canUseDOM &&
+  window.localStorage.getItem('activeTheme') || 'General Tools';