+++ packages/react/src/components/StudsCard.tsx
@@ -8,10 +8,10 @@ const StudsCard = createComponent({
     elementClass: StudsCardClass,
     react: React,
     events: {
-        onchange: 'change',
-        onclick: 'click',
-        onfocus: 'focus',
-        onblur: 'blur',
+        onChange: 'change',
+        onClick: 'click',
+        onFocus: 'focus',
+        onBlur: 'blur'
     },
 });