+++ packages/react/src/components/StudsChip.tsx
@@ -9,10 +9,11 @@ const StudsChip = createComponent({
     elementClass: StudsChipClass,
     react: React,
     events: {
-        onchange: 'change',
-        onclick: 'click',
-        onfocus: 'focus',
-        onblur: 'blur',
+        onChange: 'change',
+        onClick: 'click',
+        onFocus: 'focus',
+        onBlur: 'blur',
+        onDelete: "delete"
     },
 });