+++ packages/react/src/components/StudsInput.tsx
@@ -9,10 +9,10 @@ const StudsInput = createComponent({
     elementClass: StudsInputClass,
     react: React,
     events: {
-        onchange: 'change',
-        onclick: 'click',
-        onfocus: 'focus',
-        onblur: 'blur',
+        onChange: 'change',
+        onClick: 'click',
+        onFocus: 'focus',
+        onBlur: 'blur',
     },
 });