+++ packages/react/src/components/StudsPagination.tsx
@@ -9,10 +9,12 @@ const StudsPagination = createComponent({
     elementClass: StudsPaginationClass,
     react: React,
     events: {
-        onchange: 'change',
-        onclick: 'click',
-        onfocus: 'focus',
-        onblur: 'blur',
+        onChange: 'change',
+        onClick: 'click',
+        onFocus: 'focus',
+        onBlur: 'blur',
+        onChangePage: "changePage",
+        onChangeItemsPerPage: "changeItemsPerPage"
     },
 });